refactor(display-store): migrate vuex store to pinia in etc
