+++ apps/web/src/lib/menu/config.ts
@@ -1,5 +1,5 @@
 // Menu Ids' Rule: All menu ids are dot-delimited in depth, up to two depths.
-import type { HighlightTagType } from '@/store/modules/display/type';
+import type { HighlightTagType } from '@/store/display/type';
 
 export const MENU_ID = Object.freeze({
     WORKSPACE_HOME: 'workspace_home',
