+++ apps/web/src/schema/cost-analysis/budget/model.ts
@@ -4,7 +4,7 @@ import type {
     BudgetNotification, BudgetPlannedLimit, BudgetTimeUnit, ProviderFilter,
 } from '@/schema/cost-analysis/budget/type';
 
-import type { Currency } from '@/store/modules/display/type';
+import type { Currency } from '@/store/display/type';
 
 
 export interface BudgetModel {
