+++ apps/web/src/services/advanced/types/preferences-type.ts
@@ -1,4 +1,4 @@
-import type { Currency } from '@/store/modules/display/type';
+import type { Currency } from '@/store/display/type';
 
 export type DormancyConfig = {
     enabled: boolean;
