+++ apps/web/src/services/auth/authenticator/index.ts
@@ -2,11 +2,12 @@ import { LocalStorageAccessor } from '@cloudforet/core-lib/local-storage-accesso
 
 import { SpaceRouter } from '@/router';
 import type { AuthType } from '@/schema/identity/user/type';
-import { store } from '@/store';
 
 import { useUserWorkspaceStore } from '@/store/app-context/workspace/user-workspace-store';
+import { useDisplayStore } from '@/store/display/display-store';
 import { useDomainStore } from '@/store/domain/domain-store';
 import { useErrorStore } from '@/store/error/error-store';
+import { pinia } from '@/store/pinia';
 import { useUserStore } from '@/store/user/user-store';
 
 
@@ -16,6 +17,7 @@ abstract class Authenticator {
         const errorStore = useErrorStore();
         const domainStore = useDomainStore();
         const userStore = useUserStore();
+        const displayStore = useDisplayStore(pinia);
         await userStore.signIn({
             domainId: domainStore.state.domainId,
             credentials,
@@ -23,19 +25,20 @@ abstract class Authenticator {
             verify_code: verifyCode,
         });
         await userWorkspaceStore.load();
-        await store.dispatch('display/hideSignInErrorMessage');
+        displayStore.setIsSignInFailed(false);
         errorStore.reset();
     }
 
     static async signOut(): Promise<void> {
         const errorStore = useErrorStore();
         const userStore = useUserStore();
+        const displayStore = useDisplayStore(pinia);
         try {
             if (SpaceRouter.router) {
-                await userStore.signOut();
+                userStore.signOut();
                 const userWorkspaceStore = useUserWorkspaceStore();
                 userWorkspaceStore.reset();
-                await store.dispatch('display/hideSignInErrorMessage');
+                displayStore.setIsSignInFailed(false);
                 LocalStorageAccessor.removeItem('hideNotificationEmailModal');
                 errorStore.reset();
             }
