+++ apps/web/src/services/cost-explorer/stores/cost-analysis-page-store.ts
@@ -15,8 +15,8 @@ import type { WorkspaceModel } from '@/schema/identity/workspace/model';
 
 import { useAppContextStore } from '@/store/app-context/app-context-store';
 import { useUserWorkspaceStore } from '@/store/app-context/workspace/user-workspace-store';
+import type { Currency } from '@/store/display/type';
 import { useDomainStore } from '@/store/domain/domain-store';
-import type { Currency } from '@/store/modules/display/type';
 import { useAllReferenceStore } from '@/store/reference/all-reference-store';
 import type { CostDataSourceReferenceMap } from '@/store/reference/cost-data-source-reference-store';
